# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Bóveda de Documentos",
        "description": "API RESTful para gestión de documentos",
        "contact": {
            "email": "admin@bovedadocumentos.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/documents": {
            "get": {
                "tags": [
                    "Documentos"
                ],
                "summary": "Listar todos los documentos del usuario",
                "operationId": "0eeb1a024876b0279b451f715fb6ed7f",
                "parameters": [
                    {
                        "name": "category_id",
                        "in": "query",
                        "description": "Filtrar por categoría",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Buscar por título",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de documentos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Document"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Documentos"
                ],
                "summary": "Subir un nuevo documento",
                "operationId": "c9fef7703e9cd343c6c74a1a2fcf023e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "title",
                                    "file"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Contrato 2024"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Contrato de servicios"
                                    },
                                    "file": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "category_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "is_public": {
                                        "type": "boolean",
                                        "example": false
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Documento subido exitosamente"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/documents/{id}": {
            "get": {
                "tags": [
                    "Documentos"
                ],
                "summary": "Obtener información de un documento",
                "operationId": "c6cc59b0640a09a31ed2a7965f93d67b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Información del documento"
                    },
                    "403": {
                        "description": "No autorizado"
                    },
                    "404": {
                        "description": "Documento no encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Documentos"
                ],
                "summary": "Actualizar información de un documento",
                "operationId": "ab4a342b912782e0fe834285390ff14c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "category_id": {
                                        "type": "integer"
                                    },
                                    "is_public": {
                                        "type": "boolean"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Documento actualizado"
                    },
                    "403": {
                        "description": "No autorizado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Documentos"
                ],
                "summary": "Eliminar un documento",
                "operationId": "6b128368bb7ca99054b61fa94cd12162",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Documento eliminado"
                    },
                    "403": {
                        "description": "No autorizado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/documents/{id}/download": {
            "get": {
                "tags": [
                    "Documentos"
                ],
                "summary": "Descargar un documento",
                "operationId": "82e9bca1bbc2236748b6450b0502a5f0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Archivo descargado"
                    },
                    "403": {
                        "description": "No autorizado"
                    },
                    "404": {
                        "description": "Documento no encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/documents/{id}/share": {
            "post": {
                "tags": [
                    "Documentos"
                ],
                "summary": "Compartir un documento con otro usuario",
                "operationId": "d75d05e59ba0e7b350f1e3b78fcef966",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "user_id"
                                ],
                                "properties": {
                                    "user_id": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "permission": {
                                        "type": "string",
                                        "enum": [
                                            "view",
                                            "edit",
                                            "download"
                                        ],
                                        "example": "view"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Documento compartido"
                    },
                    "403": {
                        "description": "No autorizado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/documents/shared": {
            "get": {
                "tags": [
                    "Documentos"
                ],
                "summary": "Listar documentos compartidos conmigo",
                "operationId": "da16b4f9535ec4c33b655a7e03fd55d0",
                "responses": {
                    "200": {
                        "description": "Lista de documentos compartidos"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "Document": {
                "title": "Documento",
                "required": [
                    "id",
                    "title",
                    "filename",
                    "user_id"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "title": {
                        "type": "string",
                        "example": "Contrato 2024"
                    },
                    "description": {
                        "type": "string",
                        "example": "Contrato de servicios"
                    },
                    "filename": {
                        "type": "string",
                        "example": "abc123.pdf"
                    },
                    "original_filename": {
                        "type": "string",
                        "example": "contrato.pdf"
                    },
                    "mime_type": {
                        "type": "string",
                        "example": "application/pdf"
                    },
                    "file_size": {
                        "type": "integer",
                        "example": 1024000
                    },
                    "file_path": {
                        "type": "string",
                        "example": "documents/abc123.pdf"
                    },
                    "category_id": {
                        "type": "integer",
                        "example": 1,
                        "nullable": true
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "is_public": {
                        "type": "boolean",
                        "example": false
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}